feat: implementado nova factory e incrementado a lógica do controller

+++ src/main/java/com/outsera/csvreader/dto/ProducerIntervalDTO.java
@@ -4,7 +4,7 @@ public class ProducerIntervalDTO {
 
 	private String producer;
 
-	private Integer interval;
+	private Long interval;
 
 	private Long previousWin;
 
@@ -18,11 +18,11 @@ public class ProducerIntervalDTO {
 		this.producer = producer;
 	}
 
-	public Integer getInterval() {
+	public Long getInterval() {
 		return interval;
 	}
 
-	public void setInterval(Integer interval) {
+	public void setInterval(Long interval) {
 		this.interval = interval;
 	}
